+++ README.md
@@ -348,4 +348,11 @@
 
 - [ ] HTTP 응답 코드 설정 (200 OK, 201 Created, 404 Not Found 등)
 
+## 피드백
+- [x] 인증 처리를 여전히 여러 서비스들을 컨트롤러에서 하는 문제 
+  - [x] 비지니스 로직은 서비스 내부로 정리하기
+- [ ] SpringSecurity가 필수가 아님 
+  - [ ] 단순한 인증 인가
+  - [ ] Resolver를 컴포넌트 계층단위로 구분
+
