+++ build.gradle
@@ -27,7 +27,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation:2.7.3'
 
     // JWT 관련 의존성
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.6'
+    compileOnly 'io.jsonwebtoken:jjwt-api:0.12.6'
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.6'
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.6'
