+++ src/main/java/gift/dto/LoginResponse.java
@@ -6,4 +6,8 @@ public class LoginResponse {
     public LoginResponse(String token) {
         this.token = token;
     }
+
+    public String getToken() {
+        return token;
+    }
 }
