+++ src/main/java/gift/service/MemberService.java
@@ -1,9 +1,12 @@
 package gift.service;
 
 import gift.domain.Member;
+import gift.dto.LoginRequest;
+import gift.dto.LoginResponse;
 import gift.dto.MemberRequest;
 import gift.dto.MemberResponse;
 import gift.repository.MemberRepository;
+import gift.security.JwtTokenProvider;
 import gift.security.SecurityService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -12,11 +15,13 @@ import org.springframework.stereotype.Service;
 public class MemberService {
     private final MemberRepository memberRepository;
     private final BCryptPasswordEncoder passwordEncoder;
+    private final JwtTokenProvider jwtTokenProvider;
     private final SecurityService securityService;
 
-    public MemberService(MemberRepository memberRepository, BCryptPasswordEncoder passwordEncoder,SecurityService securityService) {
+    public MemberService(MemberRepository memberRepository, BCryptPasswordEncoder passwordEncoder, JwtTokenProvider jwtTokenProvider, SecurityService securityService) {
         this.memberRepository = memberRepository;
         this.passwordEncoder = passwordEncoder;
+        this.jwtTokenProvider = jwtTokenProvider;
         this.securityService = securityService;
 }
 
@@ -32,6 +37,16 @@ public class MemberService {
 
         String token = securityService.generateJwtToken(member);
         return new MemberResponse(token);
+    }
+
+    public LoginResponse login(LoginRequest loginRequest) {
+        Member member = findByEmail(loginRequest.getEmail());
+        if (member != null && passwordEncoder.matches(loginRequest.getPassword(), member.getPassword())) {
+            String token = jwtTokenProvider.generateToken(member);
+            return new LoginResponse(token);
+        } else {
+            return new LoginResponse("일치하는 이메일이 없거나 비밀번호가 틀렸습니다.");
+        }
     }
 
     public Member findByEmail(String email) {
